requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-dotenv
streamlit
langchain
google-generativeai
python-multipart
dotenv